refactor: make batch structure more dominant as it describes input
