+++ synth-converter/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod convert;
+pub mod batch;
 pub mod rdf;
 mod graph;
-mod parser;
