--- synth-converter/src/parser/mod.rs
@@ -1,2 +1,0 @@
-pub mod actions;
-pub mod parser;
